Fixed entity visibility calculation. It was always being computed using the air play height. CPlayAreaManager::IsVisible scroll parameter has been removed, now this method only takes into account the visibility from the play camera at the correct height.
Fighter implementation was relaying on the scroll parameter to remove planes without route (fleying or heading to player). The impl was not correct because it could cause shadows to vanish unexpectedly. It has been replaced by a simple solution, planes are removed when they are not visible for a configurable timeout.