On-access resource load and resource prefetch mechanism.
Models are no longer loaded when deserialized. They are loaded in one of these three ways:
- Just in time when first accessing (getters)/rendering the model.
- By explicitly loading it via a Load call.
- By explicitly preload an unserialized model by calling Prepare.

 Only the used models are loaded when opening a level and all are prefeched to avoid unexpected delays when playing.

A chain of PrepareResources has been implemented, this chain begins in  the playareamanager, who calls the PrepareResources of the play area elements, the player and the entity types used in entity layers. Play area elements call PrepareResources on the entity types and formation types. Entity types call prepareResource of animations, animations over animation objetcts and model animation objetcts over the models.

This mechanism is currently only used for models, but can be also used for textures, sounds, etc.

The overall result is:

- Scenario editor load time dramatically improved.
- Formation editor load time dramatically improved.
- Level load time improved in the game.

Pending work:

To be effective in the EntityEditor, the main entity selector must be replaced with a text list.
To be really effective in the game,  all systems must be kept between executions. This can be done now that the resources are not loaded when deserialized.