Fixed issue 4 (crash in netbooks) and other visual defects in low-end graphic boards.
Particle  stages were being rendered on the shadow map and somehow opengl state changes in the particle rendering were causing the intel's driver crash. The root cause has not been determined yet. It will be analyzed, untill then, Particle stages are removed from they shadow generation because the particles do not cast them.

Added support for the height fog that was missing when there is no shader support. It has been implemented as a simple fog, but it only is active if the camera is looking at 0,-1,0.

Texture levels for shadows, sky shadow and normal maps have been renumbered to fix shadow map generation when there is no shader support.

Back buffer shadow implementation was ill-implemented, it used the current game gui window viewport to deduce the maximun texture size. Now it used at most half of the computer screen, as the backbuffer size has nothing to do with the viewport size. Half of the screen is taken instead of the full screen becuase graphic boards not implementing FrameBuffer and PBuffer and very likely to be very slow.