Fixed Issue 5. Open AL sound source management refactored to work with only a few sources.
In some hosts, OpenAL supports a very reduced ammount of sources, (reduced in comparison to the currently used sources in the engine). This caused a lot of sounds to fail loading, because the sources were allocated for each sound type at load time.

Sound source management has been refactored to handle all the sounds with the minimal number of sources.
Now the SoundManager allocates a fixed number of sources (16 by default), and temporary gives them to the Sound Types.
Sound types have a configurable maximun number of channels (that is, sources) that can be active for that type, and they request at most that maximun number of sources to the SoundManager. When a channel in a sound type is not used it is  returned to the SoundManager. If a Sound type has reached its maximun of channels and a new sound is played, the oldest  sound source is reclaimed from the sound and reused in the new one.

During testing, the maximun number of concurrent channels has raised  only to 7.

Tested in desktop linux, desktop windows and Acer Aspire One. 
No problem has been detected so far.