WIP: Android port. Removed glMapBuffer[OES].
System font references in GameGUI replaced by texture fonts.

Removed glMapBuffer[OES] from OpenGLModel, it gives very little performance boost  because it is only called in buffer generation. It was the only GLES 2.0 extension used, it is known to perform poorly in some devices and in the android emulator it returns 0x2b address with no error code, causing a crash .